requirements.txt
streamlit
langchain-openai
langchain-core
PyMuPDF
pydantic>=2.0.0
python-dotenv
openai
azure-identity
azure-core
pandas
matplotlib
Pillow
numpy
